--- client/client/src/styles/forum.css
@@ -1,259 +1,0 @@
-/* Forum styles */
-.forum-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 20px;
-}
-
-.forum-header {
-    font-size: 24px;
-    font-weight: bold;
-    margin-bottom: 20px;
-}
-
-.forum-post {
-    width: 100%;
-    max-width: 800px;
-    background-color: #f9f9f9;
-    border-radius: 8px;
-    padding: 15px;
-    margin-bottom: 15px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.forum-post-title {
-    font-size: 20px;
-    font-weight: bold;
-    margin-bottom: 10px;
-}
-
-.forum-post-content {
-    font-size: 16px;
-    line-height: 1.5;
-}
-
-/* New styles for the forum component */
-.posts-container {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.post {
-    width: 100%;
-    max-width: 800px;
-    background-color: #ffffff;
-    border-radius: 8px;
-    padding: 15px;
-    margin-bottom: 15px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.post h2 {
-    font-size: 20px;
-    font-weight: bold;
-    margin-bottom: 10px;
-    color: #333;
-}
-
-.post p {
-    font-size: 16px;
-    line-height: 1.5;
-    color: #555;
-}
-
-.user-profile-image {
-    width: 30px;
-    height: 30px;
-    border-radius: 50%;
-    margin-right: 5px;
-    vertical-align: middle;
-}
-
-.username {
-    color: #0078d7;
-    cursor: pointer;
-    font-weight: bold;
-}
-
-.username:hover {
-    text-decoration: underline;
-}
-
-.tags {
-    display: flex;
-    align-items: center;
-    margin-top: 10px;
-}
-
-.tags p {
-    margin-right: 5px;
-    font-weight: bold;
-    color: #333;
-}
-
-.tags span {
-    background-color: #e0e0e0;
-    padding: 5px 10px;
-    border-radius: 15px;
-    margin-right: 5px;
-    font-size: 14px;
-    color: #333;
-}
-
-.show-comments-btn {
-    background-color: #007bff;
-    color: white;
-    padding: 10px 15px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    margin-top: 10px;
-    transition: background-color 0.3s ease;
-}
-
-.show-comments-btn:hover {
-    background-color: #0056b3;
-}
-
-.comments {
-    margin-top: 15px;
-    padding-left: 20px;
-    border-left: 2px solid #ccc;
-}
-
-.comments p {
-    font-weight: bold;
-    color: #333;
-    margin-bottom: 5px;
-}
-
-.comment {
-    margin-bottom: 10px;
-    padding-bottom: 10px;
-    border-bottom: 1px solid #eee;
-    display: flex;
-    align-items: flex-start;
-    gap: 10px;
-}
-
-.comment-user-image {
-    width: 24px;
-    height: 24px;
-    border-radius: 50%;
-    vertical-align: top;
-    margin-top: 3px;
-}
-
-.comment span {
-    font-size: 14px;
-    color: #555;
-    line-height: 1.4;
-}
-
-.profile-card {
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    background-color: white;
-    padding: 20px;
-    border-radius: 8px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-    z-index: 10;
-    width: 300px;
-    max-height: 80vh;
-    overflow-y: auto;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    border: 1px solid #ccc;
-}
-
-.close-profile-card {
-    position: absolute;
-    top: 10px;
-    right: 10px;
-    font-size: 20px;
-    cursor: pointer;
-    border: none;
-    background: none;
-}
-
-#user-profile {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 100%;
-}
-
-#user-profile img {
-    width: 80px;
-    height: 80px;
-    border-radius: 50%;
-    margin-bottom: 10px;
-    border: 2px solid #007bff;
-}
-
-#user-profile h2 {
-    font-size: 18px;
-    margin-bottom: 10px;
-    color: #2c3e50;
-    text-align: center;
-}
-
-#user-profile p {
-    font-size: 14px;
-    margin-bottom: 5px;
-    color: #555;
-    text-align: center;
-}
-
-#goToTopBtn {
-    position: fixed;
-    bottom: 20px;
-    right: 20px;
-    background-color: #007bff;
-    color: white;
-    padding: 10px 15px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    opacity: 0.7;
-    transition: opacity 0.3s ease;
-}
-
-#goToTopBtn:hover {
-    opacity: 1;
-}
-
-.comment-input-area {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    margin-top: 10px;
-}
-
-.comment-input {
-    padding: 8px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    width: calc(100% - 100px);
-}
-
-.post-comment-button {
-    padding: 8px 16px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    width: fit-content;
-}
-
-.post-comment-button:disabled {
-    background-color: #ccc;
-    cursor: not-allowed;
-}
