+++ client/package.json
@@ -18,6 +18,7 @@
     "nodemon": "^3.1.10",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-icons": "^5.5.0",
     "react-router-dom": "^7.5.3",
     "react-youtube": "^10.1.0"
   },
