+++ client/src/App.css
@@ -45,9 +45,9 @@ nav {
 nav ul {
   list-style: none;
   display: flex;
-  gap: 1.5rem;
   margin: 0;
   padding: 0;
+  gap: 10px;
 }
 
 nav li {
