+++ client/src/App.jsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { useState, useEffect } from 'react'
 import Home from './components/home.jsx';
 import Courses from './components/courses.jsx';
-import AboutUs from './components/aboutUs.jsx';
+import About from './components/about.jsx';
 import XTerminal from "./components/terminal.js";
-import './App.css';
 import Forum from './components/forum.jsx';
 import Login from './components/login.jsx';
 import Register from './components/register.jsx';
+import './App.css';
 
 function App() {
     const [currentUrl, setCurrentUrl] = useState(window.location.href);
@@ -31,36 +31,50 @@ function App() {
   
     return (
       <div>
-        <header>
-          <h1>Linux Learning Platform</h1>
-        </header>
-        <main>
-            <nav>
-                <ul>
-                    <li><a href="/home">Home</a></li>
-                    <li><a href="/courses">Courses</a></li>
-                    <li><a href="/forum">Forum</a></li>
-                    <li><a href="/aboutUs">About Us</a></li>
-                    <li><a href="/terminal">Terminal</a></li>
-                    <li><a href ="/login"></a></li>
-                    <li><a href="/register">Register</a></li>
-                </ul>
-            </nav>
-            
-            {currentUrl.includes('home') && <Home />}
-            {currentUrl.includes('aboutUs') && <AboutUs />}
-            {currentUrl.includes('courses') && <Courses />}
-            {currentUrl.includes('forum') && <Forum />} 
-            {currentUrl.includes('terminal') && <XTerminal />}
-            {currentUrl.includes('login') && <Login />}
-            {currentUrl.includes('register') && <Register />}
-        </main>
-        <footer>
-          <p>© LXLP. All Rights Reserved</p>
-          <a href="mailto:linuxlearningplatform@gmail.com">linuxlearningplatform@gmail.com</a>
-        </footer>
+      <header>
+        <h1>Linux Learning Platform</h1>
+      </header>
+      <main>
+        <nav>
+          <ul>
+            <li><a href="/home">Home</a></li>
+            <li><a href="/courses">Courses</a></li>
+            <li><a href="/forum">Forum</a></li>
+            <li><a href="/about">About</a></li>
+            <li><a href="/terminal">Terminal</a></li>
+            <li><a href="/login">Login</a></li>
+            <li><a href="/register">Register</a></li>
+          </ul>
+        </nav>
+        
+        {currentUrl.includes('home') && <Home />}
+        {currentUrl.includes('about') && <About />}
+        {currentUrl.includes('courses') && <Courses />}
+        {currentUrl.includes('forum') && <Forum />} 
+        {currentUrl.includes('terminal') && <XTerminal />}
+        {currentUrl.includes('login') && <Login />}
+        {currentUrl.includes('register') && <Register />}
+      </main>
+      <footer>
+        <p>© LXLP. All Rights Reserved</p>
+        <a href="mailto:linuxlearningplatform@gmail.com">linuxlearningplatform@gmail.com</a>
+      </footer>
       </div>
-    )
+    );
+
+    function switchStylesheet(theme) {
+      const link = document.querySelector('link[href="./App.css"]');
+      if (link) {
+        if (theme === 'dark-theme') {
+          link.href = 'App2.css';
+        } else if (theme === 'white-theme') {
+          link.href = 'App1.css';
+        } else {
+          link.href = 'App.css';
+        }
+      }
+      document.body.className = theme;
+    }
   }
   
   export default App
