+++ client/src/components/home.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import '../styles/home.css';
-import linuxLogo from '../assets/Linux.png';
+import linuxLogo from '../assets/linux.png';
 
 export default function Home() {
     return (
