+++ client/src/styles/courses.css
