+++ client/src/styles/login.css
@@ -4,7 +4,6 @@
     border-radius: 8px;
     max-width: 400px;
     margin: 2rem auto;
-    background-color: #f9f9f9;
   }
   
   .login-fieldset legend {
